tools/daily-watch/pyproject.toml
[project]
name = "daily-watchlist"
version = "1.1.0"
description = "Claude Code workflow for daily investment watchlist monitoring."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "requests>=2.28.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/Benboerba620/daily-watchlist"
Issues = "https://github.com/Benboerba620/daily-watchlist/issues"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
